requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2

# HTTP Requests & APIs
requests>=2.31.0
httpx>=0.25.0

# Environment & Configuration
python-dotenv>=1.0.0

# AI & Machine Learning
openai>=1.0.0

# Reddit API
praw>=7.7.0

# Date/Time Utilities
python-dateutil>=2.8.0

# CORS Support
fastapi-cors>=0.0.6

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Development tools (optional)
jupyter>=1.0.0
ipython>=8.0.0
notebook>=7.0.0

# Progress bars
tqdm>=4.66.0
